Index Resolver Values (#746)
