+++ .changeset/fine-teeth-visit.md
@@ -1,0 +1,5 @@
+---
+"ensindexer": minor
+---
+
+added optional indexAdditionalResolverRecords runtime flag to allow the indexing of additional resolver record values
