+++ .changeset/itchy-wolves-fall.md
@@ -1,0 +1,5 @@
+---
+"@ensnode/ensnode-schema": minor
+---
+
+addded resolver record value entities to schema
