+++ apps/ensindexer/.env.local.example
@@ -51,6 +51,7 @@ ENS_DEPLOYMENT_CHAIN=mainnet
 # Plugin Configuration
 # Identify which indexer plugins to activate (see `src/plugins` for available plugins)
 # This is a comma separated list of one or more available plugin names (case-sensitive).
+# NOTE: for subgraph-compatible indexing, the only valid configuration is `ACTIVE_PLUGINS=subgraph`
 ACTIVE_PLUGINS=subgraph,basenames,lineanames,threedns
 
 # Unknown label healing
@@ -82,6 +83,20 @@ ENSADMIN_URL=https://admin.ensnode.io
 # If this is set to true, ENSIndexer will attempt to heal subnames of addr.reverse
 # If this is not set, the default value is set to `DEFAULT_HEAL_REVERSE_ADDRESSES`.
 #
-# WARNING: Setting this to `true` results in indexed data no longer being backwards compatible with the ENS Subgraph. For full
-# data-level backwards compatibility with the ENS Subgraph, set this to `false`.
+# WARNING: Setting this to `true` results in indexed data no longer being backwards compatible with
+# the ENS Subgraph. For full data-level backwards compatibility with the ENS Subgraph, set this to `false`.
 HEAL_REVERSE_ADDRESSES=true
+
+# A feature flag to enable or disable the indexing of additional Resolver record values
+# If this is set to true, ENSIndexer will track additional Resolver record values.
+# If this is not set, the default value is set to `DEFAULT_INDEX_ADDITIONAL_RESOLVER_RECORDS` (true).
+#
+# WARNING: It is unsafe to assume that indexed resolver record values are equivalent to the
+# result of performing dynamic forward resolution via the ENS protocol, as naively retrieving resolver
+# record values from indexed data is not ENSIP-10 nor CCIP-Read compliant. At this time we do _not_
+# recommend anyone directly using the indexed resolver record values; Features are planned in the
+# ENSNode roadmap that will provide safe use of indexed resolver record values (in appropriate contexts).
+#
+# NOTE: Setting this to `true` results in indexed data being a strict _superset_ of subgraph data.
+# For exact subgraph data compatibility, set this to `false`.
+INDEX_ADDITIONAL_RESOLVER_VALUES=true
