+++ apps/ensindexer/src/api/index.ts
@@ -8,6 +8,7 @@ import { client, graphql as ponderGraphQL } from "ponder";
 
 import config from "@/config";
 import { makeApiDocumentationMiddleware } from "@/lib/api-documentation";
+import { filterSchemaExtensions } from "@/lib/filter-schema-extensions";
 import { fixContentLengthMiddleware } from "@/lib/fix-content-length-middleware";
 import {
   fetchEnsRainbowVersion,
@@ -20,6 +21,8 @@ import {
   buildGraphQLSchema as buildSubgraphGraphQLSchema,
   graphql as subgraphGraphQL,
 } from "@ensnode/ponder-subgraph";
+
+const schemaWithoutExtensions = filterSchemaExtensions(schema);
 
 const app = new Hono();
 
@@ -81,12 +84,12 @@ app.get(
 );
 
 // use ponder client support
-app.use("/sql/*", client({ db, schema }));
+app.use("/sql/*", client({ db, schema: schemaWithoutExtensions }));
 
 // use ponder middleware at `/ponder` with description injection
 app.use("/ponder", fixContentLengthMiddleware);
 app.use("/ponder", makeApiDocumentationMiddleware("/ponder"));
-app.use("/ponder", ponderGraphQL({ db, schema }));
+app.use("/ponder", ponderGraphQL({ db, schema: schemaWithoutExtensions }));
 
 // use our custom graphql middleware at /subgraph with description injection
 app.use("/subgraph", fixContentLengthMiddleware);
@@ -96,7 +99,7 @@ app.use(
   subgraphGraphQL({
     db,
     graphqlSchema: buildSubgraphGraphQLSchema({
-      schema,
+      schema: schemaWithoutExtensions,
       // provide the schema with ponder's internal metadata to power _meta
       metadataProvider: makePonderMetdataProvider({ db, publicClients }),
       // describes the polymorphic (interface) relationships in the schema
