+++ apps/ensindexer/src/config/config.schema.ts
@@ -12,6 +12,7 @@ import {
   DEFAULT_ENSADMIN_URL,
   DEFAULT_ENS_DEPLOYMENT_CHAIN,
   DEFAULT_HEAL_REVERSE_ADDRESSES,
+  DEFAULT_INDEX_ADDITIONAL_RESOLVER_RECORDS,
   DEFAULT_PORT,
   DEFAULT_RPC_RATE_LIMIT,
 } from "@/lib/lib-config";
@@ -19,168 +20,181 @@ import { uniq } from "@/lib/lib-helpers";
 import { ENSDeployments } from "@ensnode/ens-deployments";
 import { PluginName } from "@ensnode/ensnode-sdk";
 
-const parseBlockNumber = (envVarKey: string) =>
+const chainIdSchema = z.number().int().min(1);
+
+// parses an env string bool with strict requirement of 'true' or 'false'
+const makeEnvStringBoolSchema = (envVarKey: string) =>
+  z
+    .string()
+    .pipe(
+      z.enum(["true", "false"], {
+        error: `${envVarKey} must be 'true' or 'false'.`,
+      }),
+    )
+    .transform((val) => val === "true");
+
+const makeUrlSchema = (envVarKey: string) =>
+  z.url({
+    error: `${envVarKey} must be a valid URL string (e.g., http://localhost:8080 or https://example.com).`,
+  });
+
+const makeBlockNumberSchema = (envVarKey: string) =>
   z.coerce
     .number({ error: `${envVarKey} must be a positive integer.` })
     .int({ error: `${envVarKey} must be a positive integer.` })
     .min(0, { error: `${envVarKey} must be a positive integer.` })
     .optional();
 
-const parseRpcUrl = () =>
-  z.url({
-    error:
-      "RPC_URL must be a valid URL string (e.g., http://localhost:8080 or https://example.com).",
+const RpcConfigSchema = z.object({
+  url: makeUrlSchema("RPC_URL_*"),
+  maxRequestsPerSecond: z.coerce
+    .number({ error: "RPC_REQUEST_RATE_LIMIT_* must be an integer." })
+    .int({ error: "RPC_REQUEST_RATE_LIMIT_* must be an integer." })
+    .min(1, { error: "RPC_REQUEST_RATE_LIMIT_* must be at least 1." })
+    .default(DEFAULT_RPC_RATE_LIMIT),
+});
+
+const EnsDeploymentChainSchema = z
+  .enum(Object.keys(ENSDeployments) as [keyof typeof ENSDeployments], {
+    error: (issue) => {
+      return `Invalid ENS_DEPLOYMENT_CHAIN. Supported ENS deployment chains are: ${Object.keys(
+        ENSDeployments,
+      ).join(", ")}`;
+    },
+  })
+  .default(DEFAULT_ENS_DEPLOYMENT_CHAIN);
+
+const BlockrangeSchema = z
+  .object({
+    startBlock: makeBlockNumberSchema("START_BLOCK"),
+    endBlock: makeBlockNumberSchema("END_BLOCK"),
+  })
+  .refine(
+    (val) =>
+      val.startBlock === undefined || val.endBlock === undefined || val.endBlock > val.startBlock,
+    { error: "END_BLOCK must be greater than START_BLOCK." },
+  );
+
+const EnsNodePublicUrlSchema = makeUrlSchema("ENSNODE_PUBLIC_URL");
+const EnsAdminUrlSchema = makeUrlSchema("ENSADMIN_URL").default(DEFAULT_ENSADMIN_URL);
+
+const PonderDatabaseSchemaSchema = z
+  .string({
+    error: "DATABASE_SCHEMA is required.",
+  })
+  .trim()
+  .min(1, {
+    error: "DATABASE_SCHEMA is required and cannot be an empty string.",
   });
 
-const parseRpcMaxRequestsPerSecond = () =>
-  z.coerce
-    .number({ error: "RPC max requests per second must be an integer." })
-    .int({ error: "RPC max requests per second must be an integer." })
-    .min(1, { error: "RPC max requests per second must be at least 1." })
-    .default(DEFAULT_RPC_RATE_LIMIT);
-
-const parseChainConfig = () =>
-  z.object({
-    url: parseRpcUrl(),
-    maxRequestsPerSecond: parseRpcMaxRequestsPerSecond(),
-  });
-
-const parseEnsDeploymentChain = () =>
-  z
-    .enum(Object.keys(ENSDeployments) as [keyof typeof ENSDeployments], {
-      error: (issue) => {
-        return `Invalid ENS_DEPLOYMENT_CHAIN. Supported ENS deployment chains are: ${Object.keys(
-          ENSDeployments,
-        ).join(", ")}`;
-      },
-    })
-    .default(DEFAULT_ENS_DEPLOYMENT_CHAIN);
-
-const parseBlockrange = () =>
-  z
-    .object({
-      startBlock: parseBlockNumber("START_BLOCK"),
-      endBlock: parseBlockNumber("END_BLOCK"),
-    })
-    .refine(
-      (val) =>
-        val.startBlock === undefined || val.endBlock === undefined || val.endBlock > val.startBlock,
-      { error: "END_BLOCK must be greater than START_BLOCK." },
-    );
-
-const parseEnsNodePublicUrl = () =>
-  z.url({
-    error:
-      "ENSNODE_PUBLIC_URL must be a valid URL string (e.g., http://localhost:8080 or https://example.com).",
-  });
-
-const parseEnsAdminUrl = () =>
-  z
-    .url({
-      error:
-        "ENSADMIN_URL must be a valid URL string (e.g., http://localhost:8080 or https://example.com).",
-    })
-    .default(DEFAULT_ENSADMIN_URL);
-
-const parsePonderDatabaseSchema = () =>
-  z
-    .string({
-      error: "DATABASE_SCHEMA is required.",
-    })
-    .trim()
-    .min(1, {
-      error: "DATABASE_SCHEMA is required and cannot be an empty string.",
-    });
-
-const parsePlugins = () =>
-  z.coerce
-    .string()
-    .transform((val) => val.split(",").filter(Boolean))
-    .pipe(
-      z
-        .array(
-          z.enum(PluginName, {
-            error: `ACTIVE_PLUGINS must be a comma separated list with at least one valid plugin name. Valid plugins are: ${Object.values(
-              PluginName,
-            ).join(", ")}`,
-          }),
-        )
-        .min(1, {
+const PluginsSchema = z.coerce
+  .string()
+  .transform((val) => val.split(",").filter(Boolean))
+  .pipe(
+    z
+      .array(
+        z.enum(PluginName, {
           error: `ACTIVE_PLUGINS must be a comma separated list with at least one valid plugin name. Valid plugins are: ${Object.values(
             PluginName,
           ).join(", ")}`,
         }),
-    )
-    .refine((arr) => arr.length === uniq(arr).length, {
-      error: "ACTIVE_PLUGINS cannot contain duplicate values",
-    });
-
-const parseHealReverseAddresses = () =>
-  z
-    .string()
-    .pipe(
-      z.enum(["true", "false"], {
-        error: "HEAL_REVERSE_ADDRESSES must be 'true' or 'false'.",
+      )
+      .min(1, {
+        error: `ACTIVE_PLUGINS must be a comma separated list with at least one valid plugin name. Valid plugins are: ${Object.values(
+          PluginName,
+        ).join(", ")}`,
       }),
-    )
-    .transform((val) => val === "true")
-    .default(DEFAULT_HEAL_REVERSE_ADDRESSES);
-
-const parsePort = () =>
-  z.coerce
-    .number({ error: "PORT must be an integer." })
-    .int({ error: "PORT must be an integer." })
-    .min(1, { error: "PORT must be an integer between 1 and 65535." })
-    .max(65535, { error: "PORT must be an integer between 1 and 65535." })
-    .default(DEFAULT_PORT);
-
-const parseEnsRainbowEndpointUrl = () =>
-  z.url({
-    error:
-      "ENSRAINBOW_URL must be a valid URL string (e.g., http://localhost:8080 or https://example.com).",
+  )
+  .refine((arr) => arr.length === uniq(arr).length, {
+    error: "ACTIVE_PLUGINS cannot contain duplicate values",
   });
 
-const parseRpcConfigs = () =>
-  z.record(z.string().transform(Number), parseChainConfig(), {
-    error: "Chains configuration must be an object mapping numeric chain IDs to their configs.",
-  });
+const HealReverseAddressesSchema = makeEnvStringBoolSchema("HEAL_REVERSE_ADDRESSES") //
+  .default(DEFAULT_HEAL_REVERSE_ADDRESSES);
 
-const parseDatabaseUrl = () =>
-  z.union(
-    [
-      z.string().refine((url) => {
-        try {
-          if (!url.startsWith("postgresql://") && !url.startsWith("postgres://")) {
-            return false;
-          }
-          const config = parseConnectionString(url);
-          return !!(config.host && config.port && config.database);
-        } catch {
+const indexAdditionalResolverRecordsSchema = makeEnvStringBoolSchema(
+  "INDEX_ADDITIONAL_RESOLVER_RECORDS",
+).default(DEFAULT_INDEX_ADDITIONAL_RESOLVER_RECORDS);
+
+const PortSchema = z.coerce
+  .number({ error: "PORT must be an integer." })
+  .int({ error: "PORT must be an integer." })
+  .min(1, { error: "PORT must be an integer between 1 and 65535." })
+  .max(65535, { error: "PORT must be an integer between 1 and 65535." })
+  .default(DEFAULT_PORT);
+
+const EnsRainbowEndpointUrlSchema = makeUrlSchema("ENSRAINBOW_URL");
+
+const RpcConfigsSchema = z.record(
+  z.string().transform(Number).pipe(chainIdSchema),
+  RpcConfigSchema,
+  {
+    error: "Chains configuration must be an object mapping valid chain IDs to their configs.",
+  },
+);
+
+const DatabaseUrlSchema = z.union(
+  [
+    z.string().refine((url) => {
+      try {
+        if (!url.startsWith("postgresql://") && !url.startsWith("postgres://")) {
           return false;
         }
-      }),
-      z.undefined(),
-    ],
-    {
-      message:
-        "Invalid PostgreSQL connection string. Expected format: postgresql://username:password@host:port/database",
-    },
-  );
+        const config = parseConnectionString(url);
+        return !!(config.host && config.port && config.database);
+      } catch {
+        return false;
+      }
+    }),
+    z.undefined(),
+  ],
+  {
+    message:
+      "Invalid PostgreSQL connection string. Expected format: postgresql://username:password@host:port/database",
+  },
+);
+
+const derive_isSubgraphCompatible = <
+  CONFIG extends Pick<
+    ENSIndexerConfig,
+    "plugins" | "healReverseAddresses" | "indexAdditionalResolverRecords"
+  >,
+>(
+  config: CONFIG,
+): CONFIG & { isSubgraphCompatible: boolean } => {
+  // 1. only the subgraph plugin is active
+  const onlySubgraphPluginActivated =
+    config.plugins.length === 1 && config.plugins[0] === PluginName.Subgraph;
+
+  // 2. healReverseAddresses = false
+  // 3. indexAdditionalResolverRecords = false
+  const indexingBehaviorIsSubgraphCompatible =
+    !config.healReverseAddresses && !config.indexAdditionalResolverRecords;
+
+  return {
+    ...config,
+    isSubgraphCompatible: onlySubgraphPluginActivated && indexingBehaviorIsSubgraphCompatible,
+  };
+};
 
 const ENSIndexerConfigSchema = z
   .object({
-    ensDeploymentChain: parseEnsDeploymentChain(),
-    globalBlockrange: parseBlockrange(),
-    ensNodePublicUrl: parseEnsNodePublicUrl(),
-    ensAdminUrl: parseEnsAdminUrl(),
-    ponderDatabaseSchema: parsePonderDatabaseSchema(),
-    plugins: parsePlugins(),
-    healReverseAddresses: parseHealReverseAddresses(),
-    port: parsePort(),
-    ensRainbowEndpointUrl: parseEnsRainbowEndpointUrl(),
-    rpcConfigs: parseRpcConfigs(),
-    databaseUrl: parseDatabaseUrl(),
+    ensDeploymentChain: EnsDeploymentChainSchema,
+    globalBlockrange: BlockrangeSchema,
+    ensNodePublicUrl: EnsNodePublicUrlSchema,
+    ensAdminUrl: EnsAdminUrlSchema,
+    ponderDatabaseSchema: PonderDatabaseSchemaSchema,
+    plugins: PluginsSchema,
+    healReverseAddresses: HealReverseAddressesSchema,
+    indexAdditionalResolverRecords: indexAdditionalResolverRecordsSchema,
+    port: PortSchema,
+    ensRainbowEndpointUrl: EnsRainbowEndpointUrlSchema,
+    rpcConfigs: RpcConfigsSchema,
+    databaseUrl: DatabaseUrlSchema,
   })
+  // inject ENSIndexerConfig.isSubgraphCompatible
+  .transform(derive_isSubgraphCompatible)
+  // perform invariant checks
   .check(invariant_requiredDatasources)
   .check(invariant_rpcConfigsSpecifiedForIndexedChains)
   .check(invariant_globalBlockrange)
