+++ apps/ensindexer/src/lib/lib-config.ts
@@ -1,9 +1,10 @@
-import type { RpcConfigEnvironment } from "@/config/types";
+import type { ENSIndexerConfig, RpcConfigEnvironment } from "@/config/types";
 
 export const DEFAULT_RPC_RATE_LIMIT = 500;
 export const DEFAULT_ENSADMIN_URL = "https://admin.ensnode.io";
 export const DEFAULT_PORT = 42069;
 export const DEFAULT_HEAL_REVERSE_ADDRESSES = true;
+export const DEFAULT_INDEX_ADDITIONAL_RESOLVER_RECORDS = true;
 export const DEFAULT_ENS_DEPLOYMENT_CHAIN = "mainnet";
 
 /**
@@ -41,4 +42,24 @@ export function getRpcConfigsFromEnv(): Record<number, RpcConfigEnvironment> {
   });
 
   return rpcConfigs;
+}
+
+export function prettyPrintConfig(config: ENSIndexerConfig) {
+  return JSON.stringify(
+    {
+      ...config,
+      databaseUrl: "*******",
+      rpcConfigs: Object.fromEntries(
+        Object.entries(config.rpcConfigs).map(([chainId, rpcConfig]) => [
+          chainId,
+          {
+            ...rpcConfig,
+            url: "*******",
+          },
+        ]),
+      ),
+    } as ENSIndexerConfig,
+    null,
+    2,
+  );
 }
