+++ apps/ensindexer/src/plugins/threedns/handlers/ThreeDNSToken.ts
@@ -2,27 +2,36 @@ import { ponder } from "ponder:registry";
 
 import { PluginName } from "@ensnode/ensnode-sdk";
 
-import { makeResolverHandlers } from "@/handlers/Resolver";
-import { makeThreeDNSTokenHandlers } from "@/handlers/ThreeDNSToken";
+import {
+  handleABIChanged,
+  handleAddrChanged,
+  handleAddressChanged,
+  handleAuthorisationChanged,
+  handleContenthashChanged,
+  handleDNSRecordChanged,
+  handleDNSRecordDeleted,
+  handleDNSZonehashChanged,
+  handleInterfaceChanged,
+  handleNameChanged,
+  handlePubkeyChanged,
+  handleTextChanged,
+  handleVersionChanged,
+  handleZoneCreated,
+} from "@/handlers/Resolver";
+import {
+  handleNewOwner,
+  handleRegistrationCreated,
+  handleRegistrationExtended,
+  handleTransfer,
+} from "@/handlers/ThreeDNSToken";
 import { ENSIndexerPluginHandlerArgs } from "@/lib/plugin-helpers";
 import { setupRootNode } from "@/lib/subgraph-helpers";
 
-export default function ({
-  pluginName,
-  namespace,
-}: ENSIndexerPluginHandlerArgs<PluginName.ThreeDNS>) {
+export default function ({ namespace }: ENSIndexerPluginHandlerArgs<PluginName.ThreeDNS>) {
   ///
   /// ThreeDNSToken Handlers
   ///
 
-  const {
-    handleNewOwner, //
-    handleTransfer,
-    handleRegistrationCreated,
-    handleRegistrationExtended,
-  } = makeThreeDNSTokenHandlers({ pluginName });
-
-  // register each handler on each contract
   ponder.on(namespace("ThreeDNSToken:setup"), setupRootNode);
   ponder.on(namespace("ThreeDNSToken:NewOwner"), handleNewOwner);
   ponder.on(namespace("ThreeDNSToken:Transfer"), handleTransfer);
@@ -32,23 +41,6 @@ export default function ({
   ///
   /// ThreeDNS Resolver Handlers
   ///
-
-  const {
-    handleABIChanged,
-    handleAddrChanged,
-    handleAddressChanged,
-    handleAuthorisationChanged,
-    handleContenthashChanged,
-    handleDNSRecordChanged,
-    handleDNSRecordDeleted,
-    handleDNSZonehashChanged,
-    handleInterfaceChanged,
-    handleNameChanged,
-    handlePubkeyChanged,
-    handleTextChanged,
-    handleVersionChanged,
-    handleZoneCreated,
-  } = makeResolverHandlers({ pluginName });
 
   ponder.on(namespace("Resolver:AddrChanged"), handleAddrChanged);
   ponder.on(namespace("Resolver:AddressChanged"), handleAddressChanged);
