+++ apps/ensindexer/test/plugin-helpers.test.ts
@@ -1,7 +1,6 @@
 import { describe, expect, it } from "vitest";
 import { setupConfigMock } from "./utils/mockConfig";
-
-setupConfigMock();
+setupConfigMock(); // setup config mock before importing dependent modules
 
 import { makePluginNamespace } from "@/lib/plugin-helpers";
 import { PluginName } from "@ensnode/ensnode-sdk";
