+++ apps/ensindexer/test/utils/mockConfig.ts
@@ -14,6 +14,7 @@ const _defaultMockConfig = buildConfigFromEnvironment({
   plugins: "subgraph",
   ensRainbowEndpointUrl: "https://api.ensrainbow.io",
   healReverseAddresses: "true",
+  indexAdditionalResolverRecords: "true",
   port: DEFAULT_PORT.toString(),
   rpcConfigs: {
     1: {
