+++ docs/ensnode.io/src/content/docs/docs/usage/hosted-ensnode-instances.mdx
@@ -93,7 +93,7 @@ Main difference from client perspective is chain being indexed and list of plugi
     <tr>
         <td>Purpose</td>
         <td>
-            Demonstration of ENSNode's multichain ENS indexing capability. Provides a superset of the data indexed by the ENS Subgraph.
+            Demonstration of ENSNode's multichain ENS indexing capability. Provides a superset of the data indexed by the ENS Subgraph. Indexes additional Resolver record values in preparation for ENS Protocol Acceleration.
         </td>
     </tr>
     </tbody>
