+++ docs/ensnode.io/src/content/docs/ensindexer/contributing/creating-a-plugin.mdx
@@ -1,0 +1,16 @@
+---
+title: Creating a Plugin
+description: Learn how to create a plugin for ENSIndexer
+sidebar:
+  label: Creating a Plugin
+  order: 2
+---
+
+import { LinkCard } from '@astrojs/starlight/components';
+
+:::note
+This guide covers how to create a plugin for ENSIndexer.
+:::
+
+1. add new enum value to `PluignName` `packages/ensnode-utils/src/types.ts`
+2. add required datasource to config in `apps/ensindexer/src/plugins/index.ts`
