+++ terraform/main.tf
@@ -14,6 +14,7 @@ locals {
       active_plugins         = "subgraph"
       ens_deployment_chain   = "holesky"
       heal_reverse_addresses = "false"
+      index_additional_resolver_records = "false"
     }
     sepolia = {
       subdomain_prefix       = "sepolia.${local.railway_environment}"
@@ -21,6 +22,7 @@ locals {
       active_plugins         = "subgraph"
       ens_deployment_chain   = "sepolia"
       heal_reverse_addresses = "false"
+      index_additional_resolver_records = "false"
     }
     mainnet = {
       subdomain_prefix       = "mainnet.${local.railway_environment}"
@@ -28,6 +30,7 @@ locals {
       active_plugins         = "subgraph"
       ens_deployment_chain   = "mainnet"
       heal_reverse_addresses = "false"
+      index_additional_resolver_records = "false"
     }
     alpha = {
       subdomain_prefix       = "alpha.${local.railway_environment}"
@@ -35,6 +38,7 @@ locals {
       active_plugins         = "subgraph,basenames,lineanames,threedns"
       ens_deployment_chain   = "mainnet"
       heal_reverse_addresses = "true"
+      index_additional_resolver_records = "true"
     }
   }
 }
@@ -60,11 +64,12 @@ module "ensindexer" {
 
   depends_on = [null_resource.health_check]
   # Instance-specific configuration
-  subdomain_prefix       = each.value.subdomain_prefix
-  database_schema        = each.value.database_schema
-  active_plugins         = each.active_plugins
-  ens_deployment_chain   = each.ens_deployment_chain
-  heal_reverse_addresses = each.heal_reverse_addresses
+  subdomain_prefix                  = each.value.subdomain_prefix
+  database_schema                   = each.value.database_schema
+  active_plugins                    = each.active_plugins
+  ens_deployment_chain              = each.ens_deployment_chain
+  heal_reverse_addresses            = each.heal_reverse_addresses
+  index_additional_resolver_records = each.index_additional_resolver_records
 
   # Common configuration (spread operator merges the map)
   base_domain_name = local.base_domain_name
