+++ terraform/modules/ensindexer/main.tf
@@ -16,8 +16,8 @@ resource "railway_service" "ensindexer" {
   region       = var.railway_region
 }
 
-# Instance of ENSIndexer started in Ponder's "serve" mode. This mode causes the ENSIndexer instance to skip the execution of any indexing logic and instead to exclusively focus on the responsibility of serving the main "public-facing" API endpoints for the overall ENSNode deployment. 
-# This division of "start" vs "serve" responsibilities between ENSIndexer instances ensures API availability continues uninterrupted for the overall ENSNode deployment through ensnode-api even if an indexing error in ensindexer causes it to crash. 
+# Instance of ENSIndexer started in Ponder's "serve" mode. This mode causes the ENSIndexer instance to skip the execution of any indexing logic and instead to exclusively focus on the responsibility of serving the main "public-facing" API endpoints for the overall ENSNode deployment.
+# This division of "start" vs "serve" responsibilities between ENSIndexer instances ensures API availability continues uninterrupted for the overall ENSNode deployment through ensnode-api even if an indexing error in ensindexer causes it to crash.
 # The following docs explain more about Ponder's "start" vs "serve" modes: https://ponder.sh/docs/api-reference/ponder-cli#serve.
 resource "railway_service" "ensindexer_api" {
   name         = "ensindexer_api"
@@ -100,6 +100,10 @@ resource "railway_variable_collection" "ensindexer" {
       value = var.heal_reverse_addresses
     },
     {
+      name  = "INDEX_ADDITIONAL_RESOLVER_RECORDS"
+      value = var.index_additional_resolver_records
+    },
+    {
       name  = "PORT"
       value = local.application_port
     }
@@ -179,6 +183,10 @@ resource "railway_variable_collection" "ensindexer_api" {
     {
       name  = "HEAL_REVERSE_ADDRESSES"
       value = var.heal_reverse_addresses
+    },
+    {
+      name  = "INDEX_ADDITIONAL_RESOLVER_RECORDS"
+      value = var.index_additional_resolver_records
     },
     {
       name  = "PORT"
